Check for changes to scanning
This adds receiving HCI CMD packets and checking to see if
they are scan enable commands. If they are, it checks to
see if scan enable/disable and duplicate filtering settings
are different then what the current script expects. If they
are, a callback is issued so that the main script can
take any necessary action.