Support multiple manuf. data sections
Add support for a manufacturer specific data section in each of the
advertisement and the scan response packet. If there are both, they will
be returned as an array to the user.

Both fields (if they exist) will also be stored and returned if a future
advertisement doesn't have any manufacturer specific fields.